Enable Team Message Stats in default HUD
https://github.com/TeamForbiddenLLC/huds/blob/bf9cfa88c772a2930cde2219382333d7ab1780f4/inc/team_info.hud#L196

"// team message stats are used to give messages to the players about the state of things in gametypes"

This allows you to see the state of an item/state when placing your crosshair on a related object. For example, this allows you to see the name and defrost % of a frozen teammate when your crosshair is on their indicator.